[CL-3160] Clean-up and use one component for SeatInfo
