--- front/app/components/SeatInfo/SeatTrackerInfo/messages.ts
@@ -1,48 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  currentAdminSeatsTitle: {
-    id: 'app.components.SeatTrackerInfo.currentAdminSeatsTitle',
-    defaultMessage: 'Current admin seats',
-  },
-  currentCollaboratorSeatsTitle: {
-    id: 'app.components.SeatTrackerInfo.currentCollaboratorSeatsTitle',
-    defaultMessage: 'Current collaborator seats',
-  },
-  includedSeats: {
-    id: 'app.components.SeatTrackerInfo.includedSeats',
-    defaultMessage: 'Included seats',
-  },
-  additionalSeats: {
-    id: 'app.components.SeatTrackerInfo.additionalSeats',
-    defaultMessage: 'Additional seats',
-  },
-  adminMessage: {
-    id: 'app.components.SeatTrackerInfo.adminMessage',
-    defaultMessage:
-      "Your plan has {adminSeatsIncluded}. Once you’ve used all the seats, extra seats will be added under 'Additional seats'.",
-  },
-  adminSeatsIncludedSubText: {
-    id: 'app.components.SeatTrackerInfo.adminSeatsIncludedSubText',
-    defaultMessage: '{adminSeats} admin seats included',
-  },
-  collaboratorMessage: {
-    id: 'app.components.SeatTrackerInfo.collaboratorMessage',
-    defaultMessage:
-      "Your plan has {adminSeatsIncluded}, eligible for folder managers and project managers. Once you’ve used all the seats, extra seats will be added under 'Additional seats'.",
-  },
-  collaboratorIncludedSubText: {
-    id: 'app.components.SeatTrackerInfo.collaboratorIncludedSubText',
-    defaultMessage: '{projectManagerSeats} manager seats included',
-  },
-  includedAdminToolTip: {
-    id: 'app.components.SeatTrackerInfo.includedAdminToolTip',
-    defaultMessage:
-      'This shows the number of available seats for admins included in the yearly contract.',
-  },
-  includedCollaboratorToolTip: {
-    id: 'app.components.SeatTrackerInfo.includedCollaboratorToolTip',
-    defaultMessage:
-      'This shows the number of available seats for collaborators included in the yearly contract.',
-  },
-});
